Dependency cleanup
Removes some unused dependencies/features and standardizes
the format of all versions.

The most notable change is the removal of lazy_regex as it's
been replaced by a manual check that all characters are
alphanumeric. Since it's now responsible for preventing
path-traversal, some tests have also been added. Ideally
we'd have more tests covering more things, but this is a start.